refactor: moving the validation functions into the validate package
