+++ azurerm/helpers/validate/time.go
@@ -2,12 +2,28 @@ package validate
 
 import (
 	"fmt"
+	"regexp"
 	"time"
 
 	"github.com/Azure/go-autorest/autorest/date"
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/hashicorp/terraform/helper/validation"
 )
+
+func ISO8601Duration(i interface{}, k string) (warnings []string, errors []error) {
+	v, ok := i.(string)
+	if !ok {
+		errors = append(errors, fmt.Errorf("expected type of %s to be string", k))
+		return
+	}
+
+	matched, _ := regexp.MatchString(`^P([0-9]+Y)?([0-9]+M)?([0-9]+W)?([0-9]+D)?(T([0-9]+H)?([0-9]+M)?([0-9]+(\.?[0-9]+)?S)?)?$`, v)
+
+	if !matched {
+		errors = append(errors, fmt.Errorf("expected %s to be in ISO 8601 duration format, got %s", k, v))
+	}
+	return warnings, errors
+}
 
 //todo, now in terraform helper, switch over once vended
 // -> https://github.com/hashicorp/terraform/blob/master/helper/validation/validation.go#L263
