+++ azurerm/helpers/validate/time_test.go
@@ -47,7 +47,7 @@ func TestRFC3339Time(t *testing.T) {
 	}
 }
 
-func TestRfc3339DateInFutureBy(t *testing.T) {
+func TestRFC3339DateInFutureBy(t *testing.T) {
 	cases := []struct {
 		Name     string
 		Time     string
@@ -100,5 +100,45 @@ func TestRfc3339DateInFutureBy(t *testing.T) {
 				t.Fatalf("Expected RFC3339DateInFutureBy to have %d not %d errors for %q in future by %q", tc.Errors, len(errors), tc.Time, tc.Duration.String())
 			}
 		})
+	}
+}
+func TestISO8601Duration(t *testing.T) {
+	cases := []struct {
+		Value  string
+		Errors int
+	}{
+		{
+			// Date components only
+			Value:  "P1Y2M3D",
+			Errors: 0,
+		},
+		{
+			// Time components only
+			Value:  "PT7H42M3S",
+			Errors: 0,
+		},
+		{
+			// Date and time components
+			Value:  "P1Y2M3DT7H42M3S",
+			Errors: 0,
+		},
+		{
+			// Invalid prefix
+			Value:  "1Y2M3DT7H42M3S",
+			Errors: 1,
+		},
+		{
+			// Wrong order of components, i.e. invalid format
+			Value:  "PT7H42M3S1Y2M3D",
+			Errors: 1,
+		},
+	}
+
+	for _, tc := range cases {
+		_, errors := ISO8601Duration(tc.Value, "example")
+
+		if len(errors) != tc.Errors {
+			t.Fatalf("Expected ISO8601Duration to trigger '%d' errors for '%s' - got '%d'", tc.Errors, tc.Value, len(errors))
+		}
 	}
 }
