+++ azurerm/validators_test.go
@@ -153,27 +153,3 @@ func TestValidateAzureVirtualMachineTimeZone(t *testing.T) {
 		}
 	}
 }
-
-func TestValidateAzureDataLakeStoreRemoteFilePath(t *testing.T) {
-	cases := []struct {
-		Value  string
-		Errors int
-	}{
-		{
-			Value:  "bad",
-			Errors: 1,
-		},
-		{
-			Value:  "/good/file/path",
-			Errors: 0,
-		},
-	}
-
-	for _, tc := range cases {
-		_, errors := validateFilePath()(tc.Value, "unittest")
-
-		if len(errors) != tc.Errors {
-			t.Fatalf("Expected validateFilePath to trigger '%d' errors for '%s' - got '%d'", tc.Errors, tc.Value, len(errors))
-		}
-	}
-}
